Bwelly item, WITHOUT BASECODE ALTERATIONS
Realistically this is as much a tech demo of holy fuckin shit loc is scary but powerful to work with, as it is to give endosoma/stuffing enthusiasts some content.  By using an obj as the loc, you can implement handle_internal_lifeform and give it values for breathing and otherwise handling a mob inside - SHOULD protect from radiation and environmental dangers as well, though not guaranteed.  More study is needed.